users basic crud V2

+++ iobuilders-users-infrastructure/src/main/java/com/angelperez/iobuildersusers/adapter/UsersRepositoryPortAdapter.java
@@ -1,5 +1,6 @@
 package com.angelperez.iobuildersusers.adapter;
 
+import com.angelperez.iobuildersusers.common.OperationResult;
 import com.angelperez.iobuildersusers.infrastructureports.UsersRepositoryPort;
 import com.angelperez.iobuildersusers.mapper.UserEntitiesMapper;
 import com.angelperez.iobuildersusers.model.User;
@@ -23,24 +24,51 @@ public class UsersRepositoryPortAdapter implements UsersRepositoryPort {
     }
 
     @Override
-    public Mono<User> saveUser(User user) {
-        return usersRepository.save(userEntitiesMapper.toEntity(user))
-            .map(userEntitiesMapper::toDomainModel)
-            .doOnError(err -> log.error(String.format("Error saving user with id [%s]", user.getId()), err))
-            .onErrorReturn(new User());
-    }
-
-    @Override
-    public Mono<User> updateUser(User user) {
+    public Mono<OperationResult> saveUser(User user) {
         return usersRepository.findById(user.getId())
-            .map(entity -> mergeUserProperties(entity, user))
-            .doOnNext(entity -> usersRepository.save(entity))
-            .map(userEntitiesMapper::toDomainModel);
+            .hasElement()
+            .flatMap(hasValue -> {
+                if (hasValue) {
+                    return Mono.just(OperationResult.CONFLICT);
+                } else {
+                    return usersRepository.save(userEntitiesMapper.toEntity(user).setNew(true))
+                        .doOnError(err -> log.error(String.format("Error saving user with id [%s]", user.getId()), err))
+                        .map(savedEntity -> OperationResult.OK)
+                        .onErrorReturn(OperationResult.ERROR);
+                }
+            })
+            .doOnError(err -> log.error(String.format("Error searching for user with id [%s]", user.getId()), err))
+            .onErrorReturn(OperationResult.ERROR);
     }
 
     @Override
-    public Mono<Void> deleteUser(String id) {
-        return usersRepository.deleteById(id);
+    public Mono<OperationResult> updateUser(User user) {
+        return usersRepository.findById(user.getId())
+            .flatMap(entity -> usersRepository.save(mergeUserProperties(entity, user).setNew(false))
+                .doOnError(err -> log.error(String.format("Error updating user with id [%s]", user.getId()), err))
+                .map(updatedEntity -> OperationResult.OK)
+                .onErrorReturn(OperationResult.ERROR))
+            .defaultIfEmpty(OperationResult.NOT_FOUND)
+            .doOnError(err -> log.error(String.format("Error searching for user with id [%s]", user.getId()), err))
+            .onErrorReturn(OperationResult.ERROR);
+    }
+
+    @Override
+    public Mono<OperationResult> deleteUser(String id) {
+        return usersRepository.findById(id)
+            .hasElement()
+            .flatMap(hasValue -> {
+                if (!hasValue) {
+                    return Mono.just(OperationResult.NOT_FOUND);
+                } else {
+                    return usersRepository.deleteById(id)
+                        .doOnError(err -> log.error(String.format("Error deleting user with id [%s]", id), err))
+                        .map(v -> OperationResult.OK)
+                        .onErrorReturn(OperationResult.ERROR);
+                }
+            })
+            .doOnError(err -> log.error(String.format("Error searching for user with id [%s]", id), err))
+            .onErrorReturn(OperationResult.ERROR);
     }
 
     private UserEntity mergeUserProperties(UserEntity userEntity, User user) {
